requirements.txt
setuptools~=65.5.1
PyYAML~=6.0
SQLAlchemy~=1.4.44